+++ src/routes/user.route.ts
@@ -7,11 +7,12 @@ import bodyParser from "body-parser";
 import {User} from "../entities/User";
 import {VerificationHelper} from "../helpers/verficationHelper/verificationHelper";
 import {IMessageResponse} from "../interfaces/IMessageResponse";
+import bcrypt from "bcrypt";
 
 const router = Router();
 
 /**
- * @api {get} /users/:id Request for get user
+ * @api {get} /users/byId/:id Request for get user
  * @apiName get user
  * @apiGroup Users
  * @apiHeader {String} token
@@ -21,21 +22,23 @@ const router = Router();
  */
 router.get('/byId/:id', AdminMiddleware.isLogged(), async (req: Request, res: Response) => {
     const responseUser = await userFromToken(tokentSpit(req.headers["authorization"]));
-    await VerificationHelper.elementDoesNotExist(+req.params.id, res, "User");
-    if (responseUser.role.id === RoleTypes.Admin || responseUser.id === +req.params.id) {
-        try {
-            const user = await UserController.getUserById(+req.params.id);
-            res.status(200).json(user);
+    const elementDoesNotExist = await VerificationHelper.elementDoesNotExist(+req.params.id, res, "User");
+    if (elementDoesNotExist) {
+        if (responseUser.role.id === RoleTypes.Admin || responseUser.id === +req.params.id) {
+            try {
+                const user = await UserController.getUserById(+req.params.id);
+                res.status(200).json(user);
 
-        } catch (e) {
-            res.status(400).json(e);
+            } catch (e) {
+                res.status(400).json(e);
+            }
+        } else {
+            const response: IMessageResponse = {
+                Code: 401,
+                Message: "Vous n'êtes pas autorisé à consulter cet utilisateur"
+            };
+            res.status(response.Code).json(response.Message);
         }
-    } else {
-        const response: IMessageResponse = {
-            Code: 401,
-            Message: "Vous n'êtes pas autorisé à consulter cet utilisateur"
-        };
-        res.status(response.Code).json(response.Message);
     }
 });
 
@@ -50,17 +53,19 @@ router.get('/byId/:id', AdminMiddleware.isLogged(), async (req: Request, res: Re
  */
 router.get('/:role', AdminMiddleware.isAdmin, async (req: Request, res: Response) => {
     const responseUser = await userFromToken(tokentSpit(req.headers["authorization"]));
-    await VerificationHelper.elementDoesNotExist(+req.params.id, res, "User");
-    try {
-        const users = await UserController.getAllByRole(responseUser.id, +req.params.role);
-        res.status(200).json(users);
-    } catch (e) {
-        res.status(400).json(e);
+    const elementDoesNotExist = await VerificationHelper.elementDoesNotExist(+req.params.role, res, "Role");
+    if (elementDoesNotExist) {
+        try {
+            const users = await UserController.getAllByRole(responseUser.id, +req.params.role);
+            res.status(200).json(users);
+        } catch (e) {
+            res.status(400).json(e);
+        }
     }
 });
 
 /**
- * @api {put} /users/:id  Request for update user
+ * @api {put} /users/update/:id  Request for update user
  * @apiName update User
  * @apiGroup Users
  * @apiHeader {String} token admin
@@ -69,15 +74,14 @@ router.get('/:role', AdminMiddleware.isAdmin, async (req: Request, res: Response
  * @apiParam {string} firstname of user
  * @apiParam {string} lastname of user
  * @apiParam {string} email of user
- * @apiParam {string} password of user
  *
  * @apiSuccess {User} return user
  * @apiError  {string} unauthorize
  */
 router.put('/update/:id', [AdminMiddleware.isLogged(), bodyParser.json()], async (req: Request, res: Response) => {
-    const {firstname, lastname, email, password} = req.body;
-    VerificationHelper.allRequiredParam(firstname, lastname, email, password, res);
-    await VerificationHelper.elementDoesNotExist(+req.params.id, res, "User");
+    const {firstname, lastname, email} = req.body;
+    const allRequireParams = VerificationHelper.allRequiredParam(firstname, lastname, email, res);
+    const elementDoesNotExist = await VerificationHelper.elementDoesNotExist(+req.params.id, res, "User");
     const userByEmail = await userFromEmail(email);
     const userById = await UserController.getUserById(+req.params.id);
     // Si un utilisateur existe avec l'adresse email rentrée (userByEmail)
@@ -92,18 +96,20 @@ router.put('/update/:id', [AdminMiddleware.isLogged(), bodyParser.json()], async
     } else {
         const responseUser = await userFromToken(tokentSpit(req.headers["authorization"]));
         if (responseUser.role.id === RoleTypes.Admin || responseUser.id === +req.params.id) {
-            const UserSend = {
-                firstname: req.body.firstname,
-                lastname: req.body.lastname, email: req.body.email,
-                password: req.body.password
-            };
-            try {
-                const user = await UserController.updateUser(+req.params.id, UserSend as User);
-                res.status(200).json(user);
-            } catch (e) {
-                res.status(400).json(e);
+            if (elementDoesNotExist && allRequireParams) {
+                const UserSend = {
+                    id: +req.params.id,
+                    firstname: firstname,
+                    lastname: lastname,
+                    email: email
+                };
+                try {
+                    const user = await UserController.updateUser(UserSend as User);
+                    res.status(200).json(user);
+                } catch (e) {
+                    res.status(400).json(e);
+                }
             }
-
         } else {
             const response: IMessageResponse = {
                 Code: 401,
@@ -111,6 +117,91 @@ router.put('/update/:id', [AdminMiddleware.isLogged(), bodyParser.json()], async
             };
             res.status(response.Code).json(response.Message);
         }
+    }
+});
+
+/**
+ * @api {put} /users/password/:id  Request for update user password
+ * @apiName update User
+ * @apiGroup Users
+ * @apiHeader {String} token admin
+ * @apiPermission role admin
+ *
+ * @apiParam {string} lastPassword of user
+ * @apiParam {string} newPassword of user
+ * @apiParam {string} confirmation of user
+ *
+ * @apiSuccess {User} return user
+ * @apiError  {string} unauthorize
+ */
+router.put("/update/password/:id", bodyParser.json(), AdminMiddleware.isLogged(), async (req: Request, res: Response) => {
+    const {lastPassword, newPassword, confirmation} = req.body;
+    const allReqquireParams = await VerificationHelper.allRequiredParam(lastPassword, newPassword, confirmation, res);
+    const passwordCormimationGood = VerificationHelper.passwordCormimationGood(newPassword, confirmation, res);
+    const userLogged = await userFromToken(tokentSpit(req.headers["authorization"]));
+    if (!userLogged) {
+        res.status(401).json("Tu ne peux pas être la mon grand");
+        return;
+    }
+    const isCorrect: boolean = await bcrypt.compare(newPassword, userLogged.password);
+    if (isCorrect) {
+        if (allReqquireParams && passwordCormimationGood) {
+            const UserSend = {
+                id: +req.params.id,
+                password: newPassword
+            };
+            try {
+                const user = await UserController.updateUser(UserSend as User);
+                res.status(200).json(user);
+            } catch (e) {
+                res.status(400).json(e);
+            }
+        }
+    } else {
+        res.status(400).json("Le mot de passe saisit est incorrect");
+        return;
+    }
+});
+
+/**
+ * @api {post} /users/passwordForgot Request for password forgot
+ * @apiName password forgot
+ * @apiGroup Users
+ *
+ * @apiParam {string} mail of user
+ *
+ * @apiSuccess {string} return Un email vous a été envoyer.
+ * @apiError  {string} Ce mail ne correspond à aucun de nos mail
+ */
+router.post("/passwordForgot", bodyParser.json(), async (req: Request, res: Response) => {
+    const user = await userFromEmail(req.body.email);
+    if (!user) {
+        res.status(400).json("Ce mail ne correspond à aucun de nos mail");
+        return;
+    }
+    const response = await UserController.passwordForgot(user);
+    res.status(response.Code).json(response.Message);
+
+});
+
+/**
+ * @api {post} /users/resetPassword/:road  Request for reset password
+ * @apiName reset password
+ * @apiGroup Users
+ *
+ * @apiParam {password} password of user
+ * @apiParam {PasswordConfirm} PasswordConfirm of user
+ *
+ * @apiSuccess {string} return Votre mot de passe à bien été modifier
+ */
+router.post("/resetPassword/:road", [bodyParser.json(), AdminMiddleware.roadIsAccessible()], async (req: Request, res: Response) => {
+    const road = req.params.road;
+    const password = req.body.password;
+    const passwordConfirm = req.body.passwordConfirm;
+    const passwordIsgood = VerificationHelper.passwordCormimationGood(password, passwordConfirm, res);
+    if (passwordIsgood) {
+        const response = await UserController.resetPassword(road, password);
+        res.status(response.Code).json(response.Message);
     }
 });
 
@@ -124,13 +215,14 @@ router.put('/update/:id', [AdminMiddleware.isLogged(), bodyParser.json()], async
  * @apiSuccess {string} return success
  */
 router.delete('/:id', [AdminMiddleware.isAdmin()], async (req: Request, res: Response) => {
-    await VerificationHelper.elementDoesNotExist(+req.params.id, res, "User");
-    try {
-        const response = await UserController.deleteUser(+req.params.id);
-        res.status(response.Code).json(response.Message);
-    } catch (e) {
-        res.status(400).json(e);
+    const elementDoesNotExist = await VerificationHelper.elementDoesNotExist(+req.params.id, res, "User");
+    if (elementDoesNotExist) {
+        try {
+            const response = await UserController.deleteUser(+req.params.id);
+            res.status(response.Code).json(response.Message);
+        } catch (e) {
+            res.status(400).json(e);
+        }
     }
-
-})
+});
 export default router;
